Fix BackupItem class getting wrong parameters
Passing into the class the namedtuple item that has already been
created, rather than using the *item to create a new Item.
Changed the __init__() signature so that it will accept the item
passed in, and then check the optional parameters with a getattr().
This will set the optional parameters to None if not found.